Fixed a bug with not supporting statuses of ids larger than 2.147e9.

+++ qtwitter-app/src/core.cpp
@@ -59,14 +59,14 @@ Core::Core( MainWindow *parent ) :
 
   twitterapi = new TwitterAPIInterface( this );
   connect( twitterapi, SIGNAL(newEntry(TwitterAPI::SocialNetwork,QString,Entry)), this, SLOT(addEntry(TwitterAPI::SocialNetwork,QString,Entry)) );
-  connect( twitterapi, SIGNAL(deleteEntry(TwitterAPI::SocialNetwork,QString,int)), this, SLOT(deleteEntry(TwitterAPI::SocialNetwork,QString,int)) );
-  connect( twitterapi, SIGNAL(favoriteStatus(TwitterAPI::SocialNetwork,QString,int,bool)), this, SLOT(setFavorited(TwitterAPI::SocialNetwork,QString,int,bool)) );
+  connect( twitterapi, SIGNAL(deleteEntry(TwitterAPI::SocialNetwork,QString,quint64)), this, SLOT(deleteEntry(TwitterAPI::SocialNetwork,QString,quint64)) );
+  connect( twitterapi, SIGNAL(favoriteStatus(TwitterAPI::SocialNetwork,QString,quint64,bool)), this, SLOT(setFavorited(TwitterAPI::SocialNetwork,QString,quint64,bool)) );
   connect( twitterapi, SIGNAL(postDMDone(TwitterAPI::SocialNetwork,QString,TwitterAPI::ErrorCode)), this, SIGNAL(confirmDMSent(TwitterAPI::SocialNetwork,QString,TwitterAPI::ErrorCode)) );
-  connect( twitterapi, SIGNAL(deleteDMDone(TwitterAPI::SocialNetwork,QString,int,TwitterAPI::ErrorCode)), this, SLOT(deleteEntry(TwitterAPI::SocialNetwork,QString,int)) );
+  connect( twitterapi, SIGNAL(deleteDMDone(TwitterAPI::SocialNetwork,QString,quint64,TwitterAPI::ErrorCode)), this, SLOT(deleteEntry(TwitterAPI::SocialNetwork,QString,quint64)) );
   connect( twitterapi, SIGNAL(errorMessage(QString)), this, SIGNAL(errorMessage(QString)) );
   connect( twitterapi, SIGNAL(unauthorized(TwitterAPI::SocialNetwork,QString,QString)), this, SLOT(slotUnauthorized(TwitterAPI::SocialNetwork,QString,QString)) );
-  connect( twitterapi, SIGNAL(unauthorized(TwitterAPI::SocialNetwork,QString,QString,QString,int)), this, SLOT(slotUnauthorized(TwitterAPI::SocialNetwork,QString,QString,QString,int)) );
-  connect( twitterapi, SIGNAL(unauthorized(TwitterAPI::SocialNetwork,QString,QString,int,Entry::Type)), this, SLOT(slotUnauthorized(TwitterAPI::SocialNetwork,QString,QString,int,Entry::Type)) );
+  connect( twitterapi, SIGNAL(unauthorized(TwitterAPI::SocialNetwork,QString,QString,QString,quint64)), this, SLOT(slotUnauthorized(TwitterAPI::SocialNetwork,QString,QString,QString,quint64)) );
+  connect( twitterapi, SIGNAL(unauthorized(TwitterAPI::SocialNetwork,QString,QString,quint64,Entry::Type)), this, SLOT(slotUnauthorized(TwitterAPI::SocialNetwork,QString,QString,quint64,Entry::Type)) );
 
   connect( this, SIGNAL(newRequest()), SLOT(slotNewRequest()) );
   connect( twitterapi, SIGNAL(requestDone(TwitterAPI::SocialNetwork,QString,int)), this, SLOT(slotRequestDone(TwitterAPI::SocialNetwork,QString,int)) );
@@ -81,10 +81,10 @@ Core::Core( MainWindow *parent ) :
   parent->setListViewModel( statusModel );
 
   connect( statusModel, SIGNAL(openBrowser(QUrl)), this, SLOT(openBrowser(QUrl)) );
-  connect( statusModel, SIGNAL(reply(QString,int)), this, SIGNAL(addReplyString(QString,int)) );
+  connect( statusModel, SIGNAL(reply(QString,quint64)), this, SIGNAL(addReplyString(QString,quint64)) );
   connect( statusModel, SIGNAL(about()), this, SIGNAL(about()) );
-  connect( statusModel, SIGNAL(destroy(TwitterAPI::SocialNetwork,QString,int,Entry::Type)), this, SLOT(destroy(TwitterAPI::SocialNetwork,QString,int,Entry::Type)) );
-  connect( statusModel, SIGNAL(favorite(TwitterAPI::SocialNetwork,QString,int,bool)), this, SLOT(favoriteRequest(TwitterAPI::SocialNetwork,QString,int,bool)) );
+  connect( statusModel, SIGNAL(destroy(TwitterAPI::SocialNetwork,QString,quint64,Entry::Type)), this, SLOT(destroy(TwitterAPI::SocialNetwork,QString,quint64,Entry::Type)) );
+  connect( statusModel, SIGNAL(favorite(TwitterAPI::SocialNetwork,QString,quint64,bool)), this, SLOT(favoriteRequest(TwitterAPI::SocialNetwork,QString,quint64,bool)) );
   connect( statusModel, SIGNAL(postDM(TwitterAPI::SocialNetwork,QString,QString)), this, SLOT(postDMDialog(TwitterAPI::SocialNetwork,QString,QString)) );
   connect( statusModel, SIGNAL(retweet(QString)), this, SIGNAL(addRetweetString(QString)) );
   connect( statusModel, SIGNAL(markEverythingAsRead()), this, SLOT(markEverythingAsRead()) );
@@ -250,7 +250,7 @@ void Core::get()
     emit requestStarted();
 }
 
-void Core::post( TwitterAPI::SocialNetwork network, const QString &login, const QString &status, int inReplyTo )
+void Core::post( TwitterAPI::SocialNetwork network, const QString &login, const QString &status, quint64 inReplyTo )
 {
   twitterapi->postUpdate( network, login, accountsModel->account( network, login )->password, status, inReplyTo );
   emit newRequest();
@@ -258,7 +258,7 @@ void Core::post( TwitterAPI::SocialNetwork network, const QString &login, const 
   checkForNew = false;
 }
 
-void Core::destroy( TwitterAPI::SocialNetwork network, const QString &login, int id, Entry::Type type )
+void Core::destroy( TwitterAPI::SocialNetwork network, const QString &login, quint64 id, Entry::Type type )
 {
   if ( settings.value( "General/confirmTweetDeletion", true ).toBool() ) {
     QMessageBox *confirm = new QMessageBox( QMessageBox::Warning,
@@ -283,7 +283,7 @@ void Core::destroy( TwitterAPI::SocialNetwork network, const QString &login, int
   emit requestStarted();
 }
 
-void Core::favoriteRequest( TwitterAPI::SocialNetwork network, const QString &login, int id, bool favorited )
+void Core::favoriteRequest( TwitterAPI::SocialNetwork network, const QString &login, quint64 id, bool favorited )
 {
   qDebug() << "Core::favoriteRequest()";
   if ( favorited )
@@ -449,7 +449,7 @@ void Core::addEntry( TwitterAPI::SocialNetwork network, const QString &login, En
   }
 }
 
-void Core::deleteEntry( TwitterAPI::SocialNetwork network, const QString &login, int id )
+void Core::deleteEntry( TwitterAPI::SocialNetwork network, const QString &login, quint64 id )
 {
   Account *account = accountsModel->account( network, login );
   if ( statusLists.contains( *account ) ) {
@@ -458,7 +458,7 @@ void Core::deleteEntry( TwitterAPI::SocialNetwork network, const QString &login,
   }
 }
 
-void Core::setFavorited( TwitterAPI::SocialNetwork network, const QString &login, int id, bool favorited )
+void Core::setFavorited( TwitterAPI::SocialNetwork network, const QString &login, quint64 id, bool favorited )
 {
   Account *account = accountsModel->account( network, login );
   if ( statusLists.contains( *account ) ) {
@@ -491,7 +491,7 @@ void Core::slotUnauthorized( TwitterAPI::SocialNetwork network, const QString &l
   get( network, account->login, account->password );
 }
 
-void Core::slotUnauthorized( TwitterAPI::SocialNetwork network, const QString &login, const QString &password, const QString &status, int inReplyToId )
+void Core::slotUnauthorized( TwitterAPI::SocialNetwork network, const QString &login, const QString &password, const QString &status, quint64 inReplyToId )
 {
   Q_UNUSED(password);
   Account *account = accountsModel->account( network, login );
@@ -509,7 +509,7 @@ void Core::slotUnauthorized( TwitterAPI::SocialNetwork network, const QString &l
   postDM( network, account->login, screenName, text );
 }
 
-void Core::slotUnauthorized( TwitterAPI::SocialNetwork network, const QString &login, const QString &password, int destroyId, Entry::Type type )
+void Core::slotUnauthorized( TwitterAPI::SocialNetwork network, const QString &login, const QString &password, quint64 destroyId, Entry::Type type )
 {
   Q_UNUSED(password);
   Account *account = accountsModel->account( network, login );
@@ -760,7 +760,7 @@ void Core::shortenUrl( const QString &url )
     \sa post(), destroyStatus(), authDataDialog()
 */
 
-/*! \fn void Core::post( const QByteArray &status, int inReplyTo = -1 )
+/*! \fn void Core::post( const QByteArray &status, quint64 inReplyTo = -1 )
     Sends a new Status with a content given by \a status. If user's authenticaton
     data is missing, pops up an authentication dialog.
     \param status New Status's text.
