+++ qtwitter-app/src/mainwindow.h
@@ -70,10 +70,10 @@ public slots:
 signals:
   void updateStatuses();
   void openTwitPicDialog();
-  void post( TwitterAPI::SocialNetwork network, const QString &login, QString status, int inReplyTo );
+  void post( TwitterAPI::SocialNetwork network, const QString &login, QString status, quint64 inReplyTo );
   void openBrowser( QUrl address );
   void settingsDialogRequested();
-  void addReplyString( const QString& user, int inReplyTo );
+  void addReplyString( const QString& user, quint64 inReplyTo );
   void addRetweetString( QString message );
   void resizeView( int width, int oldWidth );
   void switchModel( TwitterAPI::SocialNetwork network, const QString &login );
